Added tests. Done: froster config (user, aws, nih, s3). Pending: froster config and froster config --slurm

+++ tests/test_config.py
@@ -1,9 +1,10 @@
+from froster import *
 from unittest.mock import patch
 import unittest
+import configparser
 import os
 import shutil
-import configparser
-from froster import *
+import tempfile
 import warnings
 warnings.filterwarnings("always", category=ResourceWarning)
 warnings.filterwarnings("ignore", category=ResourceWarning)
@@ -17,7 +18,6 @@ AWS_REGION = "eu-west-1"
 AWS_REGION_2 = "eu-west-2"
 AWS_PROFILE = "froster-test"
 AWS_PROFILE_2 = "froster-test-2"
-
 AWS_ACCESS_KEY_ID = os.getenv('AWS_ACCESS_KEY_ID')
 AWS_SECRET = os.getenv('AWS_SECRET')
 
@@ -27,47 +27,71 @@ SHARED_SECTION = 'SHARED'
 NIH_SECTION = 'NIH'
 S3_SECTION = 'S3'
 
-SHARED_DIR = '/tmp/shared_dir'
+S3_BUCKET_NAME = 'froster-githubactions-test'
+S3_BUCKET_NAME_2 = 'froster-githubactions-test-2'
+S3_ARCHIVE_DIR = 'froster'
+S3_ARCHIVE_DIR_2 = 'froster_2'
+S3_STORAGE_CLASS = 'DEEP_ARCHIVE'
+S3_STORAGE_CLASS_2 = 'GLACIER'
+
+SHARED_DIR = os.path.join(tempfile.gettempdir(), 'shared_dir')
 
 
 def init_froster(self):
+
     self.parser = parse_arguments()
     self.args = self.parser.parse_args()
     self.cfg = ConfigManager()
     self.arch = Archiver(self.args, self.cfg)
     self.aws = AWSBoto(self.args, self.cfg, self.arch)
 
-    # Create the data directory
+    # Create a fresh data directory
+    if os.path.exists(self.cfg.data_dir):
+        shutil.rmtree(self.cfg.data_dir)
     os.makedirs(self.cfg.data_dir, exist_ok=True, mode=0o775)
 
-    # Create the data directory
+    # Create a fresh data directory
+    if os.path.exists(self.cfg.config_dir):
+        shutil.rmtree(self.cfg.config_dir)
     os.makedirs(self.cfg.config_dir, exist_ok=True, mode=0o775)
 
-    # Create the shared directory
+    # Create a fresh shared directory
+    if os.path.exists(SHARED_DIR):
+        shutil.rmtree(SHARED_DIR)
     os.makedirs(SHARED_DIR, exist_ok=True, mode=0o775)
+
+    # Create a fresh aws directory
+    if os.path.exists(self.cfg.aws_dir):
+        shutil.rmtree(self.cfg.aws_dir)
+    os.makedirs(self.cfg.aws_dir, exist_ok=True, mode=0o775)
 
 
 def deinit_froster(self):
 
-    # Delete used directories
-    if os.path.exists(self.cfg.aws_dir):
-        shutil.rmtree(self.cfg.aws_dir)
+    if hasattr(self, 'cfg'):
+        if hasattr(self.cfg, 'aws_dir') and os.path.exists(self.cfg.aws_dir):
+            shutil.rmtree(self.cfg.aws_dir)
 
-    if os.path.exists(self.cfg.config_dir):
-        shutil.rmtree(self.cfg.config_dir)
+        if hasattr(self.cfg, 'config_dir') and os.path.exists(self.cfg.config_dir):
+            shutil.rmtree(self.cfg.config_dir)
 
-    if os.path.exists(self.cfg.data_dir):
-        shutil.rmtree(self.cfg.data_dir)
+        if hasattr(self.cfg, 'data_dir') and os.path.exists(self.cfg.data_dir):
+            shutil.rmtree(self.cfg.data_dir)
 
     if os.path.exists(SHARED_DIR):
         shutil.rmtree(SHARED_DIR)
 
     # Delete initizalized objects
-    del self.parser
-    del self.args
-    del self.cfg
-    del self.arch
-    del self.aws
+    if hasattr(self, 'parser'):
+        del self.parser
+    if hasattr(self, 'args'):
+        del self.args
+    if hasattr(self, 'cfg'):
+        del self.cfg
+    if hasattr(self, 'arch'):
+        del self.arch
+    if hasattr(self, 'aws'):
+        del self.aws
 
 
 def check_ini_file(self, ini_file, section, key, value):
@@ -90,10 +114,16 @@ class TestConfigUser(unittest.TestCase):
 
     @patch('inquirer.text', side_effect=[NAME, EMAIL])
     def test_set_user(self, mock_print, mock_input_text):
-        '''Set the user and email in the configuration file.'''
+        '''- Set the user and email in the configuration file.'''
+
+        # Check that the user is not set
+        self.assertFalse(self.cfg.user_init)
 
         # Call set_user method
         self.assertTrue(self.cfg.set_user())
+
+        # Check that the user is set
+        self.assertTrue(self.cfg.user_init)
 
         # Check that the configuration file was updated correctly
         check_ini_file(self, self.cfg.config_file,
@@ -122,10 +152,16 @@ class TestConfigAWS(unittest.TestCase):
     @patch('inquirer.list_input', side_effect=['+ Create new profile', AWS_REGION])
     @patch('inquirer.text', side_effect=[AWS_PROFILE, AWS_ACCESS_KEY_ID, AWS_SECRET])
     def test_set_aws(self, mock_print, mock_input_list, mock_input_text):
-        '''Set a new AWS profile with valid credentials.'''
+        '''- Set a new AWS profile with valid credentials.'''
+
+        # Check that the aws_init is not set
+        self.assertFalse(self.cfg.aws_init)
 
         # Call set_aws method
         self.assertTrue(self.cfg.set_aws(self.aws))
+
+        # Check that the aws_init is set
+        self.assertTrue(self.cfg.aws_init)
 
         # Check that the configuration files were updated correctly
         check_ini_file(self, self.cfg.config_file,
@@ -151,11 +187,18 @@ class TestConfigAWS(unittest.TestCase):
     @patch('inquirer.list_input', side_effect=['+ Create new profile', AWS_REGION])
     @patch('inquirer.text', side_effect=[AWS_PROFILE, 'wrong_access_key_id', 'wrong_secret'])
     def test_set_aws_invalid_credentials(self, mock_print, mock_input_list, mock_input_text):
-        '''Set a new AWS profile with invalid credentials.'''
+        '''- Set a new AWS profile with invalid credentials.'''
+
+        # Check that the aws_init is not set
+        self.assertFalse(self.cfg.aws_init)
 
         # Set valid credentials
         self.assertFalse(self.cfg.set_aws(self.aws))
 
+        # Check that the aws_init is not set
+        self.assertFalse(self.cfg.aws_init)
+
+        # Check that the configuration files were not updated
         config = configparser.ConfigParser()
         config.read(self.cfg.aws_credentials_file)
         self.assertNotIn(AWS_PROFILE, config.sections())
@@ -163,12 +206,22 @@ class TestConfigAWS(unittest.TestCase):
     @patch('inquirer.list_input', side_effect=['+ Create new profile', AWS_REGION, AWS_PROFILE, AWS_REGION])
     @patch('inquirer.text', side_effect=[AWS_PROFILE, AWS_ACCESS_KEY_ID, AWS_SECRET])
     def test_set_aws_select_profile(self, mock_print, mock_input_list, mock_input_text):
-        '''Select profile from AWS configuration file.'''
+        '''- Select profile from AWS configuration file.'''
+
+        # Check that the aws_init is not set
+        self.assertFalse(self.cfg.aws_init)
+
         # Call set_aws methon and create a new profile
         self.assertTrue(self.cfg.set_aws(self.aws))
 
+        # Check that the aws_init is set
+        self.assertTrue(self.cfg.aws_init)
+
         # Call set_aws method and select an existing profile
         self.assertTrue(self.cfg.set_aws(self.aws))
+
+        # Check that the aws_init is set
+        self.assertTrue(self.cfg.aws_init)
 
         # Check that the configuration files were updated correctly
         check_ini_file(self, self.cfg.config_file,
@@ -195,13 +248,22 @@ class TestConfigAWS(unittest.TestCase):
     @patch('inquirer.text', side_effect=[AWS_PROFILE, AWS_ACCESS_KEY_ID, AWS_SECRET, AWS_PROFILE_2, AWS_ACCESS_KEY_ID, AWS_SECRET])
     @patch('inquirer.confirm', side_effect=['y'])
     def test_set_aws_overwrite_profile(self, mock_print, mock_input_list, mock_input_text, mock_input_confirm):
-        '''Overwrite an existing AWS profile.'''
+        '''- Overwrite an existing AWS profile.'''
+
+        # Check that the aws_init is not set
+        self.assertFalse(self.cfg.aws_init)
 
         # Call set_aws method and create a new profile
         self.assertTrue(self.cfg.set_aws(self.aws))
 
+        # Check that the aws_init is set
+        self.assertTrue(self.cfg.aws_init)
+
         # Call set_aws method and overwrite the existing profile
         self.assertTrue(self.cfg.set_aws(self.aws))
+
+        # Check that the aws_init is set
+        self.assertTrue(self.cfg.aws_init)
 
         # Check that the configuration files were updated correctly
         check_ini_file(self, self.cfg.config_file,
@@ -228,12 +290,22 @@ class TestConfigAWS(unittest.TestCase):
     @patch('inquirer.text', side_effect=[AWS_PROFILE, AWS_ACCESS_KEY_ID, AWS_SECRET, AWS_PROFILE])
     @patch('inquirer.confirm', side_effect=[False])
     def test_set_aws_do_not_overwrite_profile(self, mock_print, mock_input_list, mock_input_text, mock_input_confirm):
-        '''Do not overwrite an existing AWS profile.'''
+        '''- Do not overwrite an existing AWS profile.'''
+
+        # Check that the aws_init is not set
+        self.assertFalse(self.cfg.aws_init)
+
         # Call set_aws method and create a new profile
         self.assertTrue(self.cfg.set_aws(self.aws))
 
+        # Check that the aws_init is set
+        self.assertTrue(self.cfg.aws_init)
+
         # Call set_aws method and do not overwrite the existing profile
         self.assertFalse(self.cfg.set_aws(self.aws))
+
+        # Check that the aws_init is set
+        self.assertTrue(self.cfg.aws_init)
 
         # Check that the configuration files were updated correctly
         check_ini_file(self, self.cfg.config_file,
@@ -271,7 +343,7 @@ class TestConfigShared(unittest.TestCase):
     @patch('inquirer.confirm', side_effect=[False, True])
     @patch('inquirer.prompt', return_value={'shared_dir': SHARED_DIR})
     def test_set_shared(self, mock_print, mock_input_confirm, mock_input_prompt):
-        '''Set the shared flag to False and then to True in the configuration file.'''
+        '''- Set the shared flag to False and then to True in the configuration file.'''
 
         # Call set_shared method
         self.assertTrue(self.cfg.set_shared())
@@ -297,7 +369,7 @@ class TestConfigShared(unittest.TestCase):
     @patch('inquirer.confirm', side_effect=[True, False])
     @patch('inquirer.prompt', return_value={'shared_dir': SHARED_DIR})
     def test_set_shared_do_not_move_froster_archives(self, mock_print, mock_input_confirm, mock_input_prompt):
-        '''Set the shared flag to True and do not move the froster_archives.json file.'''
+        '''- Set the shared flag to True and do not move the froster_archives.json file.'''
 
         # Create a dummy froster_archives.json file
         archive_json_file = os.path.join(
@@ -329,7 +401,7 @@ class TestConfigShared(unittest.TestCase):
     @patch('inquirer.confirm', side_effect=[True, True])
     @patch('inquirer.prompt', return_value={'shared_dir': SHARED_DIR})
     def test_set_shared_move_froster_archives(self, mock_print, mock_input_confirm, mock_input_prompt):
-        '''Set the shared flag to True and move the froster_archives.json file.'''
+        '''- Set the shared flag to True and move the froster_archives.json file.'''
 
         # Create a dummy froster_archives.json file
         archive_json_file = os.path.join(
@@ -362,7 +434,7 @@ class TestConfigShared(unittest.TestCase):
     @patch('inquirer.confirm', side_effect=[True])
     @patch('inquirer.prompt', return_value={'shared_dir': SHARED_DIR})
     def test_set_shared_froster_archives_exist(self, mock_print, mock_input_confirm, mock_input_prompt):
-        '''Set the shared flag to True when froster_archives.json file already exists in shared dir.'''
+        '''- Set the shared flag to True when froster_archives.json file already exists in shared dir.'''
 
         # Create a dummy froster_archives.json file in the shared directory
         archive_json_file_shared = os.path.join(
@@ -395,7 +467,7 @@ class TestConfigShared(unittest.TestCase):
     @patch('inquirer.confirm', side_effect=[True])
     @patch('inquirer.prompt', return_value={'shared_dir': SHARED_DIR})
     def test_set_shared_froster_shared_config_exist(self, mock_print, mock_input_confirm, mock_input_prompt):
-        '''Set the shared flag to True when froster_archives.json file already exists in shared dir.'''
+        '''- Set the shared flag to True when froster_archives.json file already exists in shared dir.'''
 
         # Create a dummy froster_archives.json file in the shared directory
         archive_json_file_shared = os.path.join(
@@ -434,10 +506,16 @@ class TestConfigNIH(unittest.TestCase):
 
     @patch('inquirer.confirm', side_effect=[True, False])
     def test_set_nih(self, mock_print, mock_input_confirm):
-        '''Set the NIH flag to True and then to False in the configuration file.'''
+        '''- Set the NIH flag to True and then to False.'''
+
+        # Check that the nih_init is not set
+        self.assertFalse(self.cfg.nih_init)
 
         # Call set_user method
         self.assertTrue(self.cfg.set_nih())
+
+        # Check that the nih_init is set
+        self.assertTrue(self.cfg.nih_init)
 
         # Check that the configuration file was updated correctly
         check_ini_file(self, self.cfg.config_file,
@@ -445,6 +523,9 @@ class TestConfigNIH(unittest.TestCase):
 
         # Call set_user method
         self.assertTrue(self.cfg.set_nih())
+
+        # Check that the nih_init is still set
+        self.assertTrue(self.cfg.nih_init)
 
         # Check that the configuration file was updated correctly
         check_ini_file(self, self.cfg.config_file,
@@ -455,39 +536,137 @@ class TestConfigNIH(unittest.TestCase):
 class TestConfigS3(unittest.TestCase):
 
     # Method executed before every test
-    def setUp(self):
+    @patch('inquirer.list_input', side_effect=['+ Create new profile', AWS_REGION])
+    @patch('inquirer.text', side_effect=[AWS_PROFILE, AWS_ACCESS_KEY_ID, AWS_SECRET])
+    @patch('builtins.print')
+    def setUp(self, mock_print, mock_input_list, mock_input_text):
+
         init_froster(self)
+
+        # Set valid credentials
+        self.assertTrue(self.cfg.set_aws(self.aws))
+
+        # Get the buckets list
+        s3_buckets = self.aws.get_buckets()
+
+        # Delete the buckets if they exists
+        if S3_BUCKET_NAME in s3_buckets:
+            self.aws.s3_client.delete_bucket(Bucket=S3_BUCKET_NAME)
+        if S3_BUCKET_NAME_2 in s3_buckets:
+            self.aws.s3_client.delete_bucket(Bucket=S3_BUCKET_NAME_2)
 
     # Method executed after every test
     def tearDown(self):
+
+        # Get the buckets list
+        s3_buckets = self.aws.get_buckets()
+
+        # Delete the buckets if they exists
+        if S3_BUCKET_NAME in s3_buckets:
+            self.aws.s3_client.delete_bucket(Bucket=S3_BUCKET_NAME)
+        if S3_BUCKET_NAME_2 in s3_buckets:
+            self.aws.s3_client.delete_bucket(Bucket=S3_BUCKET_NAME_2)
+
         deinit_froster(self)
 
     def test_set_s3_aws_not_init(self, mock_print):
-        ''' set_s3 method returns False if AWS credentials are not set.'''
+        '''-  set_s3 method returns False if AWS credentials are not set.'''
 
-        self.assertFalse(self.cfg.set_s3(self.aws))
+        # Mock the aws_init variable
+        self.cfg.aws_init = False
 
-    def test_set_s3_invalid_aws_credentials(self, mock_print):
-        '''set_s3 method returns False if AWS credentials are invalid.'''
-
-        # Set variables as if AWS credentials were set
-        self.cfg.aws_init = True
-        self.cfg.aws_profile = AWS_PROFILE
-
-        print("Check set_s3", flush=True)
         # Call set_s3 method
         self.assertFalse(self.cfg.set_s3(self.aws))
 
-    # @patch('inquirer.list_input', side_effect=['+ Create new profile', AWS_REGION, '+ Create new bucket', 'DEEP_ARCHIVE'])
-    # @patch('inquirer.text', side_effect=[AWS_PROFILE, AWS_ACCESS_KEY_ID, AWS_SECRET, 'froster-githubactions-test', 'froster'])
-    # def test_set_s3(self, mock_input_list, mock_input_text):
-    #     '''Set a new S3 bucket.'''
+        # Check that the aws_init is not set
+        self.assertFalse(self.cfg.aws_init)
 
-    #     # Set valid credentials
-    #     self.cfg.set_aws(self.aws)
+        # Check that the s3_init is not set
+        self.assertFalse(self.cfg.s3_init)
 
-    #     # Call set_s3 method
-    #     self.assertTrue(self.cfg.set_s3(self.aws))
+        # Restore the aws_init variable
+        self.cfg.aws_init = True
+
+    def test_set_s3_invalid_aws_credentials(self, mock_print):
+        '''- set_s3 method returns False if AWS credentials are invalid.'''
+
+        # Manually change to a profile not set
+        self.cfg.aws_profile = AWS_PROFILE_2
+
+        # Check that the aws_init is set
+        self.assertTrue(self.cfg.aws_init)
+
+        # Check that the s3_init is not set
+        self.assertFalse(self.cfg.s3_init)
+
+        # Call set_s3 method
+        self.assertFalse(self.cfg.set_s3(self.aws))
+
+        # Check that the aws_init is still set
+        self.assertTrue(self.cfg.aws_init)
+
+        # Check that the s3_init is not set
+        self.assertFalse(self.cfg.s3_init)
+
+        # Restore the changed profile
+        self.cfg.aws_profile = AWS_PROFILE
+
+    @patch('inquirer.list_input', side_effect=['+ Create new bucket', S3_STORAGE_CLASS])
+    @patch('inquirer.text', side_effect=[S3_BUCKET_NAME, S3_ARCHIVE_DIR])
+    def test_set_s3(self, mock_print, mock_input_list, mock_input_text):
+        '''- Set a new S3 bucket.'''
+
+        # Call set_s3 method
+        self.assertTrue(self.cfg.set_s3(self.aws))
+
+        # Check that the configuration files were updated correctly
+        check_ini_file(self, self.cfg.config_file,
+                       S3_SECTION, 'bucket_name', S3_BUCKET_NAME)
+
+        check_ini_file(self, self.cfg.config_file,
+                       S3_SECTION, 'archive_dir', S3_ARCHIVE_DIR)
+
+        check_ini_file(self, self.cfg.config_file,
+                       S3_SECTION, 'storage_class', S3_STORAGE_CLASS)
+
+        # Check that the s3_init is set
+        self.assertTrue(self.cfg.s3_init)
+
+        # Get the buckets list
+        s3_buckets = self.aws.get_buckets()
+
+        # Check the bucket was created
+        self.assertIn(S3_BUCKET_NAME, s3_buckets)
+
+    @patch('inquirer.list_input', side_effect=['+ Create new bucket', S3_STORAGE_CLASS, S3_BUCKET_NAME, S3_STORAGE_CLASS_2])
+    @patch('inquirer.text', side_effect=[S3_BUCKET_NAME, S3_ARCHIVE_DIR, S3_ARCHIVE_DIR_2])
+    def test_set_s3_select_bucket(self, mock_print, mock_input_list, mock_input_text):
+        '''- Select S3 bucket.'''
+
+        # Call set_s3 method
+        self.assertTrue(self.cfg.set_s3(self.aws))
+
+        # Call set_s3 method
+        self.assertTrue(self.cfg.set_s3(self.aws))
+
+        # Check that the configuration files were updated correctly
+        check_ini_file(self, self.cfg.config_file,
+                       S3_SECTION, 'bucket_name', S3_BUCKET_NAME)
+
+        check_ini_file(self, self.cfg.config_file,
+                       S3_SECTION, 'archive_dir', S3_ARCHIVE_DIR_2)
+
+        check_ini_file(self, self.cfg.config_file,
+                       S3_SECTION, 'storage_class', S3_STORAGE_CLASS_2)
+
+        # Delete the bucket
+        self.aws.s3_client.delete_bucket(Bucket=S3_BUCKET_NAME)
+
+        # Get the buckets list
+        s3_buckets = self.aws.get_buckets()
+
+        # Check the bucket was created
+        self.assertNotIn(S3_BUCKET_NAME, s3_buckets)
 
 
 if __name__ == '__main__':
@@ -501,7 +680,6 @@ if __name__ == '__main__':
         # suite.addTest(unittest.TestLoader().loadTestsFromTestCase(TestConfigShared))
         # suite.addTests(unittest.TestLoader().loadTestsFromTestCase(TestConfigNIH))
         # suite.addTests(unittest.TestLoader().loadTestsFromTestCase(TestConfigS3))
-        suite.addTest(TestConfigS3('test_set_s3'))
+        # suite.addTest(TestConfigS3('test_set_s3_select_bucket'))
         runner = unittest.TextTestRunner(verbosity=2)
         runner.run(suite)
-
